pet_app/requirements.txt
fastapi
uvicorn
pillow
python-multipart
torch
torchvision